% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_password.R
\name{change_password}
\alias{change_password}
\title{Change User Password}
\usage{
change_password(old_password, new_password, token)
}
\arguments{
\item{old_password}{A character string representing the user's current password.}

\item{new_password}{A character string representing the user's new password.}

\item{token}{A character string representing the user's access token.}
}
\value{
A message indicating the success of the password change, or an error if it fails.
}
\description{
Change User Password
}
