% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_forgot_password.R
\name{confirm_forgot_password}
\alias{confirm_forgot_password}
\title{Confirm Forgot Password}
\usage{
confirm_forgot_password(client_id, username, confirmation_code, new_password)
}
\arguments{
\item{client_id}{A character string representing the app client ID.}

\item{username}{A character string for the user's username.}

\item{confirmation_code}{A character string for the confirmation code sent to the user.}

\item{new_password}{A character string for the user's new password.}
}
\value{
Nothing (displays a message on success or raises an error if it fails).
}
\description{
This function confirms the forgot password process by setting a new password for a user.
}
