% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_sign_up.R
\name{confirm_sign_up_user}
\alias{confirm_sign_up_user}
\title{Confirm User Sign Up}
\usage{
confirm_sign_up_user(client_id, userpool, username, verification_code)
}
\arguments{
\item{client_id}{A character string representing the app client ID.}

\item{userpool}{A character string representing the user pool ID.}

\item{username}{A character string for the user's username.}

\item{verification_code}{A character string for the verification code sent to the user.}
}
\value{
Nothing (displays a message on success or raises an error if it fails).
}
\description{
This function confirms the sign-up process for a user in AWS Cognito.
}
