% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forgot_password.R
\name{forgot_password}
\alias{forgot_password}
\title{Initiate Forgot Password Request}
\usage{
forgot_password(client_id, username)
}
\arguments{
\item{client_id}{A character string representing the app client ID.}

\item{username}{A character string for the user's username.}
}
\value{
Nothing (displays a message on success or raises an error if it fails).
}
\description{
Initiate Forgot Password Request
}
