% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_user_list_from_cognito}
\alias{get_user_list_from_cognito}
\title{Get User List from Cognito}
\usage{
get_user_list_from_cognito(cognito_client, userpool)
}
\arguments{
\item{cognito_client}{The Cognito client.}

\item{userpool}{A character string representing the user pool ID.}
}
\value{
A list containing usernames and email addresses.
}
\description{
Retrieve the list of usernames and email addresses from Cognito.
}
