% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_email}
\alias{is_valid_email}
\title{Check if an email address is valid.}
\usage{
is_valid_email(email)
}
\arguments{
\item{email}{The email address to be validated.}
}
\value{
TRUE if the email is valid, otherwise an error is thrown.
}
\description{
This function validates the format of an email address.
}
