% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_token}
\alias{is_valid_token}
\title{Check if an access token is valid.}
\usage{
is_valid_token(token)
}
\arguments{
\item{token}{The access token to be validated.}
}
\value{
TRUE if the token is valid, otherwise an error is thrown.
}
\description{
This function validates the format of an access token.
}
