% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_up.R
\name{sign_up}
\alias{sign_up}
\title{Sign up a new user to AWS Cognito}
\usage{
sign_up(client_id, email, username, password, ...)
}
\arguments{
\item{client_id}{A character string representing the app client ID.}

\item{email}{A character string for the user's email address.}

\item{username}{A character string for the user's username.}

\item{password}{A character string for the user's password.}

\item{...}{Additional named user attributes for Cognito (e.g., `phone_number`).}
}
\value{
A list with sign-up confirmation details if successful; throws an error otherwise.
}
\description{
Sign up a new user to AWS Cognito
}
