% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_in.R
\name{sign_in}
\alias{sign_in}
\title{Sign in to AWS Cognito}
\usage{
sign_in(client_id, username, password, authflow = "USER_PASSWORD_AUTH")
}
\arguments{
\item{client_id}{A character string representing the app client ID.}

\item{username}{A character string for the user's username.}

\item{password}{A character string for the user's password.}

\item{authflow}{A character string specifying the authentication flow (default: "USER_PASSWORD_AUTH").}
}
\value{
A list with user information and tokens if successful; throws an error otherwise.
}
\description{
Sign in to AWS Cognito
}
